refactor(hooks): extract useScrollProgress, use in LoadingBar and BackToTopButton
Part of #1471. Adds src/hooks/ and a shared RAF-throttled scroll hook,
replacing the duplicated ticking/requestAnimationFrame scroll listener
in LoadingBar and BackToTopButton. Behavior preserved exactly, including
LoadingBar's deferred listener attachment until its intro animation ends.
